language update delte remaining

+++ Application/Operations/Language/Commands/CreateLanguage/CreateLanguageCommand.cs
@@ -3,13 +3,13 @@ using Domain.Events;
 using MediatR;
 
 namespace Application.CommandQueries.Language.Commands.CreateLanguage;
-public record CreateLanguageCommand : IRequest<int>
+public record CreateLanguageCommand : IRequest<short>
 {
     public string? Title { get; init; }
     public string? Code { get; init; }
 }
 
-public class CreateLanguageCommandHandler : IRequestHandler<CreateLanguageCommand, int>
+public class CreateLanguageCommandHandler : IRequestHandler<CreateLanguageCommand, short>
 {
     private readonly IApplicationDbContext _context;
 
@@ -18,7 +18,7 @@ public class CreateLanguageCommandHandler : IRequestHandler<CreateLanguageComman
         _context = context;
     }
 
-    public async Task<int> Handle(CreateLanguageCommand request, CancellationToken cancellationToken)
+    public async Task<short> Handle(CreateLanguageCommand request, CancellationToken cancellationToken)
     {
         var entity = new Domain.Entities.Language
         {
